media-source/MediaSource-duration.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>MediaSource#duration</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var contents = {'/media/white.webm': 'video/webm; codecs="vorbis,vp8"',
                '/media/white.mp4' : 'video/mp4'};

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
test(function() {
    var mediaSource = new MediaSource();
    assert_equals(mediaSource.duration, NaN);
    assert_throws({name : 'InvalidAccessError'},
                  function() {
                      mediaSource.duration = -1;
                  }, 'Set invalid duration');
    assert_throws({name : 'InvalidAccessError'},
                  function() {
                      mediaSource.duration = NaN;
                  }, 'Set invalid duration');

    assert_throws({name : 'InvalidStateError'},
                  function() {
                      mediaSource.duration = 1;
                  }, 'Not opened yet');
}, 'MediaSource#duration : initial state');
function testDuration(file, mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            mediaSource.addEventListener('sourceopen', function(e) {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                sourceBuffer.addEventListener('updatestart', function(e) {
                    t.step(function() {
                        assert_throws({name : 'InvalidStateError'},
                                      function() { mediaSource.duration = 1; });
                    });
                    t.done();
                });
                sourceBuffer.appendBuffer(data);
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'MediaSource#duration (' + mime + '): update during SourceBuffer#updating is true');
}
for(var file in contents) {
    testDuration(file, contents[file]);
}
</script>
</body>
</html>
